trying to get combined CMA-ES grasp opt to work in untangle scenario
struggling to paramterize/set hyper-params in a way that works.
I might try a new parameterization where each possible binary grasp is
represented as a category and x[0:n_possibilities] represents logits of
that distribution. That should let me bias against the current config
and the [0, 0] config in a principle manner, and it means I don't need
to bound those variables