using weld eqs to improve orientation control in threading
seems to be working better
also realized I need to use the homotopy check more carefully in
planning. We now penalize with delta to desired signature. We can
mention this as a good technique in the paper -- planning with a cost.
It's not obvious how to compute a cost between two signatures.